// src/generated/models/Dga_classificationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dga_classificationsdga_classification_level = {
  1: 'L1',
  2: 'L2',
  3: 'L3',
  4: 'GLCode'
} as const;
export type Dga_classificationsdga_classification_level = keyof typeof Dga_classificationsdga_classification_level;
export const Dga_classificationsdga_expense_type = {
  1: 'Capex',
  2: 'Opex'
} as const;
export type Dga_classificationsdga_expense_type = keyof typeof Dga_classificationsdga_expense_type;
export const Dga_classificationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dga_classificationsstatecode = keyof typeof Dga_classificationsstatecode;
export const Dga_classificationsstatuscode = {
  1: 'Draft',
  2: 'Inactive',
  776140001: 'Publish'
} as const;
export type Dga_classificationsstatuscode = keyof typeof Dga_classificationsstatuscode;

export interface Dga_classificationsBase {
  dga_account_group?: string;
  dga_arabic_name?: string;
  dga_classification_level?: Dga_classificationsdga_classification_level;
  dga_classificationid: string;
  dga_description?: string;
  dga_ebs_account_code?: string;
  dga_expense_type?: Dga_classificationsdga_expense_type;
  dga_fusion_account_code?: string;
  dga_name: string;
  "dga_parent_classification@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "OwningBusinessUnit@odata.bind"?: string;
  statecode: Dga_classificationsstatecode;
  statuscode?: Dga_classificationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dga_classifications extends Dga_classificationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dga_classification_levelname?: string;
  dga_expense_typename?: string;
  dga_parent_classificationname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dga_parent_classification?: object;
  _dga_parent_classification_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}